@@ -9,8 +9,8 @@
   },
   "author": "https://github.com/orgs/WebJamApps/teams/handlersandkickers",
   "engines": {
-    "node": ">=18.12.0 <19",
-    "npm": ">=8.19.2 <9"
+    "node": ">=18.14.1 <19",
+    "npm": ">=9.3.1 <10"
   },
   "files": [
     "webpack.config.js",
@@ -30,146 +30,130 @@
     "cc": "codeclimate analyze -f html > codeclimate.html",
     "start": "webpack serve --node-env=development --env --inline",
     "test:stylelint": "stylelint \"static/**/*.scss\"",
-    "test:lint-fix": "npm run test:stylelint && eslint . --ext .js,.ts,.tsx --fix",
+    "test:lint-fix": "eslint . --ext .js,.ts,.tsx --fix",
     "test:lint": "npm run test:stylelint && eslint . --ext .js,.ts,.tsx",
-    "test": "npm run test:lint && npm run test:jest",
-    "test:local": "npm run test:lint-fix && npm run test:jest && npm run cc",
-    "test:jest": "rm -rf coverage && jest --config jest.config.json --no-cache --runInBand",
+    "test": "npm run test:lint && npm run test:jest --",
+    "test:local": "npm run test:lint-fix && npm run test:unit && npm run cc",
+    "test:jest": "jest --config jest.config.json --runInBand",
+    "test:unit": "rm -rf coverage && npm run test:jest --",
     "cleanup": "rm -rf yarn.lock && rm -rf package-lock.json && rm -rf node_modules && npm cache clean --force",
     "cleaninstall": "npm install -g npm@latest && npm run cleanup && npm install",
     "postinstall": "npm run build:prod"
   },
   "dependencies": {
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "@hot-loader/react-dom": "^17.0.2",
-    "@material-ui/core": "^4.12.4",
-    "@mui/icons-material": "^5.10.9",
-    "@mui/material": "^5.10.12",
-    "@mui/x-data-grid": "^5.17.9",
-    "@mui/x-date-pickers": "^5.0.6",
-    "@popperjs/core": "^2.9.2",
-    "@react-oauth/google": "^0.2.8",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@tinymce/tinymce-react": "^4.2.0",
-    "@types/core-js": "^2.5.4",
-    "@types/domutils": "^1.7.2",
-    "@types/ejs": "^3.0.1",
-    "@types/enzyme": "^3.10.5",
-    "@types/faker": "^4.1.12",
-    "@types/google.maps": "^3.50.0",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.9",
+    "@mui/material": "^5.11.10",
+    "@mui/x-data-grid": "^5.17.24",
+    "@mui/x-date-pickers": "^5.0.19",
+    "@popperjs/core": "^2.11.6",
+    "@react-oauth/google": "^0.7.1",
+    "@tinymce/tinymce-react": "^4.3.0",
+    "@types/core-js": "^2.5.5",
+    "@types/domutils": "^1.7.4",
+    "@types/ejs": "^3.1.2",
+    "@types/faker": "5.5.9",
+    "@types/google.maps": "^3.52.0",
     "@types/html-webpack-plugin": "^3.2.6",
-    "@types/jest": "^29.0.0",
-    "@types/mui-datatables": "^3.7.9",
-    "@types/node": "^16.11.56",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-resize-detector": "^4.2.0",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/react-slick": "^0.23.4",
-    "@types/react-test-renderer": "^17.0.1",
-    "@types/redux-logger": "^3.0.7",
-    "@types/sanitize-html": "^1.22.0",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.14.0",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-slick": "^0.23.10",
+    "@types/react-test-renderer": "^18.0.0",
+    "@types/redux-logger": "^3.0.9",
+    "@types/sanitize-html": "^2.8.0",
     "@types/socketcluster-client": "^16.0.0",
-    "@types/superagent": "^4.1.7",
-    "@types/use-persisted-state": "^0.3.0",
-    "@types/webpack-bundle-analyzer": "^4.4.2",
+    "@types/superagent": "^4.1.16",
+    "@types/webpack-bundle-analyzer": "^4.6.0",
     "@types/webpack-env": "1.18.0",
-    "@wojtekmaj/enzyme-adapter-react-17": "latest",
-    "axios": "^1.1.2",
-    "bootstrap": "^5.2.0",
+    "axios": "^1.3.3",
+    "bootstrap": "^5.2.3",
     "copy-webpack-plugin": "^11.0.0",
-    "core-js": "^3.8.0",
+    "core-js": "^3.28.0",
     "crypto-browserify": "^3.12.0",
-    "css-loader": "^6.7.1",
-    "date-fns": "^2.28.0",
-    "dotenv": "^16.0.2",
+    "css-loader": "^6.7.3",
+    "date-fns": "^2.29.3",
+    "decode-uri-component": "^0.4.1",
+    "dotenv": "^16.0.3",
     "expose-loader": "^4.0.0",
-    "express": "^4.18.1",
+    "express": "^4.18.2",
     "express-sslify": "^1.2.0",
-    "faker": "^4.1.0",
+    "faker": "5.5.3",
     "file-loader": "^6.2.0",
-    "html-loader": "^4.1.0",
-    "html-react-parser": "^3.0.4",
+    "html-loader": "^4.2.0",
+    "html-react-parser": "^3.0.9",
     "html-webpack-plugin": "^5.5.0",
-    "jest-environment-jsdom": "^29.0.1",
-    "jquery": "^3.6.1",
-    "jsonwebtoken": "^8.5.1",
-    "mini-css-extract-plugin": "^1.3.1",
+    "jquery": "^3.6.3",
+    "jwt-simple": "^0.5.6",
+    "mini-css-extract-plugin": "^2.7.2",
     "moment": "^2.29.4",
-    "mui-datatables": "4.2.2",
     "popper.js": "^1.16.1",
     "process": "^0.11.10",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hot-loader": "^4.13.0",
-    "react-notifications-component": "^3.1.0",
-    "react-player": "^2.10.0",
-    "react-redux": "^8.0.2",
-    "react-resize-detector": "^7.1.2",
-    "react-router-dom": "^5.2.1",
-    "react-share": "^4.4.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-notifications-component": "^4.0.1",
+    "react-player": "^2.11.2",
+    "react-redux": "^8.0.5",
+    "react-resize-detector": "^8.0.4",
+    "react-router-dom": "^6.8.1",
+    "react-share": "^4.4.1",
     "react-slick": "^0.29.0",
-    "react-test-renderer": "^17.0.2",
-    "react-timekeeper": "^2.2.1",
-    "redux": "^4.2.0",
+    "react-test-renderer": "^18.2.0",
+    "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.1",
+    "redux-thunk": "^2.4.2",
     "resize-observer-polyfill": "^1.5.1",
-    "sass": "^1.35.2",
-    "sass-loader": "^13.0.2",
+    "sass": "^1.58.3",
+    "sass-loader": "^13.2.0",
     "slick-carousel": "^1.8.1",
-    "snyk": "^1.996.0",
+    "snyk": "^1.1064.0",
     "socketcluster-client": "^16.0.3",
-    "source-map-loader": "^4.0.0",
+    "source-map-loader": "^4.0.1",
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.3.1",
-    "stylelint-config-recommended-scss": "^5.0.2",
-    "superagent": "^8.0.0",
-    "terser-webpack-plugin": "^5.3.3",
-    "ts-loader": "^9.3.1",
-    "typescript": "^4.8.2",
+    "stylelint-config-recommended-scss": "^9.0.1",
+    "superagent": "^8.0.9",
+    "terser-webpack-plugin": "^5.3.6",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.5",
     "url-loader": "^4.1.1",
     "url-search-params-polyfill": "^8.1.1",
-    "use-persisted-state": "^0.3.3",
-    "util": "^0.12.3",
-    "webpack": "^5.7.0",
-    "webpack-bundle-analyzer": "^4.6.1",
-    "webpack-cli": "^4.10.0"
+    "util": "^0.12.5",
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.8.0",
+    "webpack-cli": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@types/redux-mock-store": "latest",
-    "@typescript-eslint/eslint-plugin": "^5.36.1",
-    "@typescript-eslint/parser": "^5.36.1",
-    "babel-loader": "^8.2.3",
-    "enzyme": "latest",
-    "eslint": "^8.23.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.52.0",
+    "@typescript-eslint/parser": "^5.52.0",
+    "eslint": "^8.34.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-plugin-es": "^4.1.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.0.1",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-json": "^3.1.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7.31.0",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-security": "^1.5.0",
-    "identity-obj-proxy": "latest",
-    "jest": "^29.0.1",
-    "redux-mock-store": "latest",
-    "stylelint": "^14.2.0",
-    "stylelint-config-standard": "^24.0.0",
-    "stylelint-scss": "^4.1.0",
-    "ts-jest": "29.0.0-next.0",
-    "webpack-dev-server": "^4.8.1"
+    "eslint-plugin-security": "^1.7.1",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.4.3",
+    "jest-environment-jsdom": "^29.4.3",
+    "stylelint": "^15.2.0",
+    "stylelint-config-standard": "^30.0.1",
+    "stylelint-scss": "^4.4.0",
+    "ts-jest": "29.0.5",
+    "webpack-dev-server": "^4.11.1"
   },
   "snyk": true
 }
