@@ -1,17 +1,22 @@
 import { CodeResponse, useGoogleLogin } from '@react-oauth/google';
-import type { Dispatch } from 'react';
+import { useContext } from 'react';
 import { Button } from '@mui/material';
+import { AuthContext, Iauth } from 'src/providers/Auth.provider';
 import utils from './utils';
 
-export const loginConfig = (dispatch: Dispatch<unknown>) => ({
-  onSuccess: (codeResponse: Omit<CodeResponse, 'error' | 'error_description' | 'error_uri'>) => utils.responseGoogleLogin(codeResponse, dispatch),
+export const loginConfig = (auth:Iauth, setAuth: (args0:Iauth)=>void) => ({
+  onSuccess: (codeResponse: Omit<CodeResponse, 'error' | 'error_description' | 'error_uri'>) => {
+    utils.responseGoogleLogin(codeResponse, auth, setAuth);
+  },
   onError: () => { console.log('Google login failed'); return false; },
   flow: 'auth-code',
 });
 
-export function GoogleButtons(props: { type: string, index: number, dispatch: Dispatch<unknown> }): JSX.Element {
-  const { type, index, dispatch } = props;
-  const login = useGoogleLogin(loginConfig(dispatch) as Record<string, unknown>);
+interface IgoogleButtonsProps { type: string, index: number}
+export function GoogleButtons(props: IgoogleButtonsProps): JSX.Element {
+  const { setAuth, auth } = useContext(AuthContext);
+  const { type, index } = props;
+  const login = useGoogleLogin(loginConfig(auth, setAuth) as Record<string, unknown>);
   if (type === 'login') {
     return (
       <div key={index} className="menu-item googleLogin">
@@ -29,7 +34,12 @@ export function GoogleButtons(props: { type: string, index: number, dispatch: Di
     );
   } return (
     <div key={index} className="menu-item googleLogout">
-      <Button className="logoutButton" variant="contained" size="small" onClick={() => { utils.responseGoogleLogout(dispatch); }}>
+      <Button
+        className="logoutButton"
+        variant="contained"
+        size="small"
+        onClick={() => { utils.responseGoogleLogout(setAuth); }}
+      >
         Logout
       </Button>
     </div>
