@@ -1,7 +1,7 @@
 import { CodeResponse, googleLogout } from '@react-oauth/google';
-import type { Dispatch } from 'react';
+import { defaultAuth, Iauth } from 'src/providers/Auth.provider';
 import 'react-notifications-component/dist/theme.css';
-import jwt from 'jsonwebtoken';
+import jwt from 'jwt-simple';
 import superagent from 'superagent';
 import commonUtils from 'src/lib/commonUtils';
 
@@ -12,24 +12,26 @@ export interface GoogleBody {
   state(): string,
 }
 
-const setUserRedux = async (
-  dispatch: (...args: any) => void,
+const setUserAuth = async (
+  setAuth: (arg0: Iauth) => void,
   token: string,
   userId: string | undefined,
 ) => {
-  const user = await superagent.get(`${process.env.BackendUrl}/user/${userId}`)
+  const { body } = await superagent.get(`${process.env.BackendUrl}/user/${userId}`)
     .set('Accept', 'application/json').set('Authorization', `Bearer ${token}`);
-  dispatch({ type: 'SET_USER', data: user.body, token });
+  setAuth({
+    error: '', isAuthenticated: true, token, user: body,
+  });
 };
 
-const setUser = async (dispatch: Dispatch<unknown>, token:string): Promise<void> => {
-  const { sub } = jwt.verify(token, process.env.HashString as string) as jwt.JwtPayload;
-  await setUserRedux(dispatch, token, sub);
+const setUser = async (auth: Iauth, setAuth: (args0: Iauth) => void, token: string): Promise<void> => {
+  const { sub } = jwt.decode(token, process.env.HashString as string);
+  await setUserAuth(setAuth, token, sub);
 };
 
 const authenticate = async (
   googleBody: GoogleBody,
-): Promise<{ token:string, email:string }> => {
+): Promise<{ token: string, email: string }> => {
   const { body } = await superagent.post(`${process.env.BackendUrl}/user/auth/google`)
     .set({ Accept: 'application/json' }).send(googleBody);
   return body;
@@ -42,7 +44,8 @@ const makeState = () => () => {
 
 const responseGoogleLogin = async (
   response: Omit<CodeResponse, 'error' | 'error_description' | 'error_uri'>,
-  dispatch: Dispatch<unknown>,
+  auth: Iauth,
+  setAuth: (arg0: Iauth) => void,
 ): Promise<void> => {
   try {
     const uri = window.location.href;
@@ -50,24 +53,25 @@ const responseGoogleLogin = async (
     const body = {
       clientId: process.env.GoogleClientId,
       redirectUri: !baseUri.includes('localhost')
-    && process.env.NODE_ENV === 'production' ? `https://${baseUri}`
+        && process.env.NODE_ENV === 'production' ? `https://${baseUri}`
         : `http://${baseUri}`,
       code: `${response.code}`,
       state: makeState(),
     };
     const { token } = await authenticate(body);
-    await setUser(dispatch, token);
+    await setUser(auth, setAuth, token);
   } catch (e) {
+    console.log(e);
     const eMessage = (e as Error).message;
     commonUtils.notify('Failed to authenticate', eMessage, 'danger');
+    setAuth({ ...defaultAuth, error: eMessage });
   }
 };
 
-const responseGoogleLogout = async (dispatch: Dispatch<unknown>): Promise<void> => {
-  dispatch({ type: 'LOGOUT' });
+const responseGoogleLogout = async (setAuth: (arg0: Iauth) => void): Promise<void> => {
   googleLogout();
-  await commonUtils.delay(2);
-  window.location.assign('/');
+  setAuth(defaultAuth);
+  window.location.reload();
 };
 
 export default {
