@@ -1,17 +1,5 @@
 import { Footer } from './Footer';
-
-function HeaderSection() {
-  return (
-    <div id="header" className="material-header home-header">
-      <div id="ohaflogo" className="headercontent">
-        <img alt="ohaflogo" src="../static/imgs/webjamicon7.png" className="home-header-image" />
-      </div>
-      <div className="headercontent header-text-card">
-        <h3 className="header-text" style={{ marginTop: 0 }}>Web Jam LLC</h3>
-      </div>
-    </div>
-  );
-}
+import { HeaderSection } from './HeaderSection';
 
 export function MainPanel({ children, onClick, onKeyPress }:any) {
   return (
