@@ -1,40 +1,70 @@
-/* eslint-disable react/sort-comp */
-import React, { Dispatch } from 'react';
-import { withRouter, RouteComponentProps } from 'react-router-dom';
-import { connect } from 'react-redux';
-import type { Auth } from 'src/redux/mapStoreToProps';
-import mapStoreToATemplateProps from 'src/redux/mapStoreToAppTemplateProps';
-import { PageHost } from './PageHost';
+import { SetStateAction, useState } from 'react';
+import { DrawerContainer } from './DrawerContainer';
+import { MainPanel } from './MainPanel';
 
-export interface AppTemplateProps extends RouteComponentProps {
-  heartBeat: string;
-  userCount: number;
-  auth: Auth;
-  dispatch: Dispatch<unknown>;
-  children: JSX.Element;
+export const makeDrawerClass = (menuOpen: boolean) => {
+  const className = `home-sidebar ${menuOpen ? 'open' : 'close'} drawer-container`;
+  return className;
+};
+
+export const makeHandleClose = (setMenuOpen: (value: SetStateAction<boolean>) => void) => () => setMenuOpen(false);
+
+export function toggleMobileMenu(menuOpen: boolean, setMenuOpen: (arg0: boolean) => void): void {
+  const mO = !menuOpen;
+  setMenuOpen(mO);
 }
 
-export class AppTemplate extends React.Component<AppTemplateProps, unknown> {
-  // eslint-disable-next-line @typescript-eslint/no-useless-constructor
-  constructor(props: AppTemplateProps) {
-    super(props);
-  }
+export const makeOnClick = (
+  menuOpen: boolean,
+  setMenuOpen: (value: SetStateAction<boolean>) => void,
+) => () => toggleMobileMenu(menuOpen, setMenuOpen);
 
-  render(): JSX.Element {
-    const {
-      children, userCount, heartBeat, auth, location, dispatch,
-    } = this.props;
-    return (
-      <PageHost
-        children={children}
-        userCount={userCount}
-        heartBeat={heartBeat}
-        auth={auth}
-        location={location}
-        dispatch={dispatch}
+export function handleKeyMenu(
+  e: { key: string; },
+  menuOpen: boolean,
+  setMenuOpen: (arg0: boolean) => void,
+): void {
+  if (e.key === 'Enter') toggleMobileMenu(menuOpen, setMenuOpen);
+}
+
+export const makeOnKeyPress = (
+  menuOpen: boolean,
+  setMenuOpen: (value: SetStateAction<boolean>) => void,
+) => (evt: { key: string; }) => handleKeyMenu(evt, menuOpen, setMenuOpen);
+
+export function handleEscapePress(e: { key: string; }, setMenuOpen: (arg0: boolean) => void): (void) {
+  if (e.key === 'Escape') setMenuOpen(false);
+}
+
+export const makeHandleKeyPress = (
+  setMenuOpen: (value: SetStateAction<boolean>) => void,
+) => (evt: { key: string; }) => handleEscapePress(evt, setMenuOpen);
+
+interface IpageHostProps {
+  userCount?: number, heartBeat?: string,
+  children: React.ReactNode,
+}
+export function AppTemplate(props: IpageHostProps) {
+  const {
+    // userCount, heartBeat,
+    children,
+  } = props;
+  const [menuOpen, setMenuOpen] = useState(false);
+  const handleClose = makeHandleClose(setMenuOpen);
+  const onClick = makeOnClick(menuOpen, setMenuOpen);
+  const onKeyPress = makeOnKeyPress(menuOpen, setMenuOpen);
+  const handleKeyPress = makeHandleKeyPress(setMenuOpen);
+  return (
+    <div className="page-host">
+      <DrawerContainer
+        handleKeyPress={handleKeyPress}
+        className={makeDrawerClass(menuOpen)}
+        // userCount={userCount}
+        // heartBeat={heartBeat}
+        handleClose={handleClose}
       />
-    );
-  }
+      <MainPanel children={children} onClick={onClick} onKeyPress={onKeyPress} />
+    </div>
+  );
 }
 
-export default withRouter(connect(mapStoreToATemplateProps, null)(AppTemplate));
