@@ -1,35 +1,34 @@
 
+import { StrictMode } from 'react';
 import { Provider } from 'react-redux';
-import { render } from 'react-dom';
-import { hot } from 'react-hot-loader';
+import { createRoot } from 'react-dom/client';
 import { GoogleOAuthProvider } from '@react-oauth/google';
 import { PersistGate } from 'redux-persist/integration/react';
-import ConnectedApp from './App/index';
+import { App } from './App/index';
 import store from './redux/store/index';
 import { DataProvider } from './providers/Data.provider';
-import { EditorProvider } from './providers/Editor.provider';
+import { AuthProvider } from './providers/Auth.provider';
 import '../static/styles.scss';
 
+const root = createRoot(document.getElementById('root') as HTMLElement);
 function Main() {
   return (
     <GoogleOAuthProvider clientId={process.env.GoogleClientId || ''}>
-      <DataProvider>
-        <EditorProvider>
+      <AuthProvider>
+        <DataProvider>
           <Provider store={store.store}>
             <PersistGate loading={null} persistor={store.persistor}>
-              <ConnectedApp />
+              <App />
             </PersistGate>
           </Provider>
-        </EditorProvider>
-      </DataProvider>
+        </DataProvider>
+      </AuthProvider>
     </GoogleOAuthProvider>
   );
 }
 
-const HotMain = hot(module)(Main);
-
 const renderMain = (): void => {
-  render(<HotMain />, document.getElementById('root'));
+  root.render(<StrictMode><Main /></StrictMode>);
 };
 
 renderMain();
