@@ -6,7 +6,7 @@ function WideAboutUs(): JSX.Element {
     <div className="widescreenHomepage">
       <div className="container-fluid">
         <div className="row">
-          <div className="col-md-6" style={{ top: '0', paddingRight: '6px' }}>
+          <div className="col-md-6" style={{ top: '0', paddingRight: '6px', paddingLeft: '10px' }}>
             <p style={{ marginTop: '32px', marginBottom: '40px', fontSize: '18px' }}>
               Web Jam LLC helps businesses bring their web technologies into the 21st century! Let us
               help make your website mobile-friendly and grow your business by utilizing the power of online services.
