@@ -1,76 +1,66 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { BrowserRouter, RouteComponentProps } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import type { ImenuItem } from 'src/App/AppTemplate/menuConfig';
-import { continueMenuItem, SideMenuItem } from 'src/App/AppTemplate/SideMenuItem';
+import { checkIsAllowed, ContinueMenuItem, SideMenuItem } from 'src/App/AppTemplate/SideMenuItem';
 import commonUtils from 'src/lib/commonUtils';
-import type { Auth } from 'src/redux/mapStoreToProps';
 
 describe('SideMenuItem', () => {
   it('is defined', () => {
     expect(SideMenuItem).toBeDefined();
   });
-  it('returns null for auth item when user is incorrect type', () => {
+  it('renders MakeLink for /music', () => {
     commonUtils.getUserRoles = jest.fn(() => ['tester']);
     const props = {
       index: 1,
-      auth: { isAuthenticated: true, user: { userType: 'joker' } } as Auth,
-      location: { pathname: '/music' } as RouteComponentProps['location'],
-      menu: { auth: true, link: '/dashboard' } as ImenuItem,
-      dispatch: jest.fn(),
-      handleClose: jest.fn(),
-    };
-    const smi = renderer.create(<SideMenuItem {...props} />).toJSON();
-    expect(smi).toBeNull();
-  });
-  it('returns MakeLink for /music', () => {
-    commonUtils.getUserRoles = jest.fn(() => ['tester']);
-    const props = {
-      index: 1,
-      auth: { isAuthenticated: false, user: { userType: 'joker' } } as Auth,
-      location: { pathname: '/music' } as RouteComponentProps['location'],
+      auth: { isAuthenticated: false, user: { userType: 'joker' } } as any,
       menu: { auth: false, link: '/music' } as ImenuItem,
-      dispatch: jest.fn(),
       handleClose: jest.fn(),
     };
-    const smi: any = renderer.create(<BrowserRouter><SideMenuItem {...props} /></BrowserRouter>).toJSON();
-    expect(smi.props.className).toBe('menu-item');
+    const smi = renderer.create(<BrowserRouter><SideMenuItem {...props} /></BrowserRouter>).root;
+    expect(smi).toBeDefined();
   });
   it('returns MakeLink for /music when Web Jam LLC', () => {
-    window.location.assign = jest.fn();
     commonUtils.getUserRoles = jest.fn(() => ['tester']);
     const props = {
       index: 1,
-      auth: { isAuthenticated: false, user: { userType: 'joker' } } as Auth,
-      location: { pathname: '/music' } as RouteComponentProps['location'],
+      auth: { isAuthenticated: false, user: { userType: 'joker' } } as any,
+      location: { pathname: '/music' },
       menu: { auth: false, link: '/', name: 'Web Jam LLC' } as ImenuItem,
-      dispatch: jest.fn(),
       handleClose: jest.fn(),
     };
     const smi: any = renderer.create(<BrowserRouter><SideMenuItem {...props} /></BrowserRouter>);
     expect(smi.toJSON().props.className).toBe('menu-item');
     smi.root.findByProps({ type: 'Link' }).props.handleClose();
-    expect(window.location.assign).toHaveBeenCalledWith('/');
+    expect(props.handleClose).toHaveBeenCalled();
   });
-  it('continueMenuItem returns googleLogout', () => {
-    const result: any = continueMenuItem({ type: 'googleLogout' } as ImenuItem, 1, { isAuthenticated: true } as Auth, '', jest.fn(), jest.fn());
-    expect(result.key).toBe('googleLogout');
-  });
-  it('clears storage when clicking on menu items from joshandmariamusic', () => {
-    localStorage.clear = jest.fn();
-    sessionStorage.clear = jest.fn();
+  it('ContinueMenuItem returns googleLogout', () => {
     const props = {
+      menu: { type: 'googleLogout' } as ImenuItem,
       index: 1,
-      auth: { isAuthenticated: false, user: { userType: 'joker' } } as Auth,
-      location: { pathname: '/' } as RouteComponentProps['location'],
-      menu: {
-        nav: 'jam', auth: false, link: '/songs', name: 'Songs',
-      } as ImenuItem,
-      dispatch: jest.fn(),
+      auth: { isAuthenticated: true } as any,
+      pathname: '',
       handleClose: jest.fn(),
     };
-    const smi: any = renderer.create(<BrowserRouter><SideMenuItem {...props} /></BrowserRouter>).root;
-    const result = smi.findByProps({ type: 'Link' }).props.handleClose();
-    expect(result).toBe('cleared');
+    const result: any = renderer.create(<ContinueMenuItem {...props} />).toJSON();
+    expect(result.props.className.includes('googleLogout')).toBe(true);
+  });
+  it('ContinueMenuItem when music', () => {
+    const props = {
+      menu: { type: 'link', link: '/music' } as ImenuItem,
+      index: 1,
+      auth: { isAuthenticated: true } as any,
+      pathname: '/music',
+      handleClose: jest.fn(),
+    };
+    const result: any = renderer.create(<BrowserRouter><ContinueMenuItem {...props} /></BrowserRouter>).toJSON();
+    expect(result.children[0].props.href).toBe('/music');
+  });
+  it('checkIsAllowed return false if item requires auth and userType is not allowed', () => {
+    const menu: any = { auth: true };
+    const auth: any = { isAuthenticated: true, user: { userType: 'tester' } };
+    const userRoles = ['admin'];
+    const isAllowed = checkIsAllowed(menu, auth, userRoles);
+    expect(isAllowed).toBe(false);
   });
 });
