@@ -1,6 +1,8 @@
 import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
-import { AppTemplate } from 'src/App/AppTemplate';
+import {
+  AppTemplate, handleEscapePress, handleKeyMenu, makeDrawerClass,
+} from 'src/App/AppTemplate';
 
 describe('AppTemplate', () => {
   it('renders the component', () => {
@@ -19,5 +21,18 @@ describe('AppTemplate', () => {
       <BrowserRouter><AppTemplate {...props}><div id="test-page" /></AppTemplate></BrowserRouter>,
     ).toJSON();
     expect(at).toMatchSnapshot();
+  });
+  it('makeDrawerClass', () => {
+    expect(makeDrawerClass(true)).toBe('home-sidebar open drawer-container');
+  });
+  it('handleEscapePress', () => {
+    const setMenuOpen = jest.fn();
+    handleEscapePress({ key: 'Escape' }, setMenuOpen);
+    expect(setMenuOpen).toHaveBeenCalledWith(false);
+  });
+  it('handleKeyMenu', () => {
+    const setMenuOpen = jest.fn();
+    handleKeyMenu({ key: 'Enter' }, false, setMenuOpen);
+    expect(setMenuOpen).toHaveBeenCalledWith(true);
   });
 });
