@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import {
-  DataProvider, getGigsDef, IGig, ISong, makeGetGigs, setGigsDef, setSongsDef,
+  DataProvider, getGigsDef, Igig, Isong, makeGetGigs, setGigsDef, setSongsDef, getPicsDef, getSongsDef, setPicsDef,
 } from 'src/providers/Data.provider';
 import fetchSongs from 'src/providers/fetchSongs';
 import fetchGigs from 'src/providers/fetchGigs';
@@ -14,13 +14,23 @@ describe('DataProvider', () => {
     expect(screen.getByText('Test Div Here')).toBeInTheDocument();
   });
   it('setGigsDef', () => {
-    expect(setGigsDef([] as IGig[])).toBeUndefined();
+    expect(setGigsDef([] as Igig[])).toBeUndefined();
   });
   it('setSongsDef', () => {
-    expect(setSongsDef([] as ISong[])).toBeUndefined();
+    expect(setSongsDef([] as Isong[])).toBeUndefined();
   });
   it('getGigsDef', () => {
     expect(getGigsDef()).toBe(true);
+  });
+  it('getSongsDef', async () => {
+    const result = await getSongsDef();
+    expect(result.length).toBe(0);
+  });
+  it('getPicsDef', () => {
+    expect(getPicsDef()).toBe(true);
+  });
+  it('setPicsDef', () => {
+    expect(setPicsDef(null)).toBe(undefined);
   });
   it('make getGigs and runs it', () => {
     fetchGigs.getGigs = jest.fn();
