@@ -1,5 +1,5 @@
 import scc from 'socketcluster-client';
-import fetchGigs, { defaultGig } from 'src/providers/fetchGigs';
+import fetchGigs from 'src/providers/fetchGigs';
 
 describe('fetchGigs', () => {
   it('getGigs runs successfully', () => {
@@ -10,10 +10,5 @@ describe('fetchGigs', () => {
     scc.create = jest.fn(() => { throw new Error('failed'); });
     const setGigs = jest.fn();
     expect(fetchGigs.getGigs(setGigs)).toBe(false);
-  });
-  it('validateGigsArr', () => {
-    const setGigs = jest.fn();
-    fetchGigs.validateGigsArr({ value: [defaultGig, defaultGig] }, setGigs);
-    expect(setGigs).toHaveBeenCalled();
   });
 });
