@@ -1,29 +1,20 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import mapStoreToProps from '../../src/redux/mapStoreToProps';
-import mapStoreToAppTemplateProps from '../../src/redux/mapStoreToAppTemplateProps';
+import mapStoreToProps from 'src/redux/mapStoreToProps';
+import mapStoreToAppTemplateProps from 'src/redux/mapStoreToAppTemplateProps';
 
 describe('mapStoreToProps', () => {
   it('returns', () => {
     const store:any = {
-      images: { images: [{}] },
-      sc: {},
-      auth: {},
-      tour: {},
-      showTable: {},
-      image: {},
+      sc: { scc: {} },
     };
     const newStore = mapStoreToProps(store);
-    expect(newStore.images.length).toBe(1);
+    expect(newStore.scc).toBeDefined();
   });
   it('returns for AppTemplate', () => {
     const store:any = {
-      images: { images: [{}] },
-      sc: {},
-      auth: {},
-      tour: {},
-      showTable: {},
+      sc: { userCount: 1 },
     };
-    const newStore = (mapStoreToAppTemplateProps(store) as { images:any[] });
-    expect(newStore.images.length).toBe(1);
+    const newStore = (mapStoreToAppTemplateProps(store));
+    expect(newStore.userCount).toBe(1);
   });
 });
